+++ cpp/CMakeLists.txt
@@ -54,7 +54,7 @@ add_executable(adaptive_hermite_refinement lib/test-triangle.cpp)
 target_include_directories(adaptive_hermite_refinement PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(adaptive_hermite_refinement argparse mdspan)
 
-add_library(src-lib OBJECT Naive.cpp Naive.h HermiteRunner.cpp HermiteRunner.h)
+add_library(src-lib OBJECT Naive.cpp Naive.hpp HermiteRunner.cpp HermiteRunner.hpp)
 target_include_directories(src-lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(src-lib mdspan fftw-cpp cnpy)
