+++ cpp/HermiteRunner.cpp
@@ -2,6 +2,6 @@
 // Created by Luka on 4/28/2023.
 //
 
-#include "HermiteRunner.h"
+#include "HermiteRunner.hpp"
 
 ahr::HermiteRunner::HermiteRunner(std::ostream &out) : out(out) {}
