+++ cpp/bench/timing.hpp
