+++ cpp/constants.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "typedefs.h"
+#include "typedefs.hpp"
 
 namespace ahr {
     using std::numbers::pi;
